camera: introduce environment module
This introduces an environment module for the camera to control sound
and triggers.
