+++ src/meson.build
@@ -106,6 +106,7 @@ common_sources = [
   'trx/game/camera/box_camera.c',
   'trx/game/camera/cinematic.c',
   'trx/game/camera/common.c',
+  'trx/game/camera/environment.c',
   'trx/game/camera/fixed.c',
   'trx/game/camera/photo_mode.c',
   'trx/game/camera/vars.c',
