+++ src/trx/game/camera.h
@@ -4,6 +4,7 @@
 #include <trx/game/camera/common.h>
 #include <trx/game/camera/const.h>
 #include <trx/game/camera/enum.h>
+#include <trx/game/camera/environment.h>
 #include <trx/game/camera/fixed.h>
 #include <trx/game/camera/photo_mode.h>
 #include <trx/game/camera/types.h>
