+++ src/trx/game/camera/common.h
@@ -2,11 +2,8 @@
 
 #include <trx/config/types.h>
 #include <trx/game/camera/types.h>
-#include <trx/game/rooms/types.h>
 
 void Camera_Update(void);
-void Camera_UpdateMicPosition(void);
-void Camera_EnsureEnvironment(void);
 void Camera_MoveManual(void);
 void Camera_Apply(void);
 bool Camera_IsChunky(void);
@@ -15,7 +12,6 @@ void Camera_Initialise(void);
 void Camera_ResetPosition(void);
 void Camera_Reset(void);
 void Camera_ClampInterpResult(void);
-void Camera_RefreshFromTrigger(const TRIGGER *trigger);
 
 void Camera_RegisterStrategy(CAMERA_MODE mode, CAMERA_STRATEGY strategy);
