+++ src/trx/game/camera/environment.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <trx/game/rooms/types.h>
+
+void Camera_RefreshFromTrigger(const TRIGGER *trigger);
+void Camera_EnsureEnvironment(void);
+void Camera_UpdateMicPosition(void);
